Move off deprecated Criteria and Update
